fix stop job bugs and add session stop operation when kill a job

+++ fate_flow/utils/job_utils.py
@@ -20,12 +20,14 @@ import json
 import operator
 import os
 import subprocess
+import sys
 import threading
 import typing
 import uuid
 
 import psutil
 
+from arch.api import session
 from arch.api.utils import file_utils
 from arch.api.utils.core import current_timestamp
 from arch.api.utils.core import json_loads, json_dumps
@@ -316,7 +318,7 @@ def wait_child_process(signum, frame):
             raise
 
 
-def kill_process(pid):
+def kill_process(pid, only_child=False):
     try:
         if not pid:
             return False
@@ -327,11 +329,43 @@ def kill_process(pid):
         for child in p.children(recursive=True):
             if check_job_process(child.pid):
                 child.kill()
-        if check_job_process(p.pid):
-            p.kill()
+        if not only_child:
+            if check_job_process(p.pid):
+                p.kill()
         return True
     except Exception as e:
         raise e
+
+
+def stop_executor(task):
+    task_dir = os.path.join(get_job_directory(job_id=task.f_job_id), task.f_role, task.f_party_id,
+                            task.f_component_name)
+    os.makedirs(task_dir, exist_ok=True)
+    kill_path = os.path.join(task_dir, 'kill')
+    f = open(kill_path, 'w')
+    f.close()
+    return True
+
+
+def onsignal_term(signum, frame):
+    try:
+        session.stop()
+        sys.exit(1)
+    except Exception as e:
+        pass
+
+
+def task_killed_detector(job_id, role, party_id, component_name):
+    kill_path = os.path.join(get_job_directory(job_id), str(role), str(party_id), component_name, 'kill')
+    if os.path.exists(kill_path):
+        session.stop()
+        with open(kill_path.replace('kill', 'pid'), 'r') as f:
+            pid = f.read()
+        while True:
+            kill_process(int(pid), only_child=True)
+            sys.exit(1)
+            kill_process(int(pid), only_child=False)
+    threading.Timer(0.25, task_killed_detector, args=[job_id, role, party_id, component_name]).start()
 
 
 def gen_all_party_key(all_party):
